refactor db
query bamboo in background where possible
