+++ Cargo.toml
@@ -26,3 +26,4 @@ toml = "0.5"
 base64 = "0.11"
 chrono = { version = "0.4", features = ["serde"] }
 itertools = "0.8"
+parking_lot = "0.10"
