+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod github;
 pub mod github_bot;
 pub mod http;
 pub mod issue;
+pub mod local_state;
 pub mod matrix;
 pub mod matrix_bot;
 pub mod project_info;
